requirements.txt
numpy >=1.8
scipy >=1.5
pytorch >=1.7
matplotlib >= 3.3